requirements.txt
cython==0.29.36
gym==0.21.0
matplotlib==3.5.2
mujoco_py==2.0.2.5
numpy==1.18.5
PyYAML==6.0
scipy==1.8.1
seaborn==0.11.2
termcolor==2.2.0
torch==1.11.0
pandas==1.1.5
